feat: adicionadas mais rotas relacionadas a criação de contas e a compras feitas online

+++ src/main/kotlin/models/SaleProduct.kt
@@ -4,9 +4,9 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class SaleProduct (
-    val id: Long,
-    val idProduct: Long,
-    val idSale: Long,
+    val id: Int,
+    val idProduct: Int,
+    val idSale: Int,
     val quantity: Long
 ) {
